pyproject.toml
[tool.poetry]
name = "bobpience_config"
version = "0.1.0"
description = ""
authors = ["NohSungwoo <shtjddn0817@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2"
black = {extras = ["d"], version = "^24.8.0"}
isort = "^5.13.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
drf-spectacular = "^0.27.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.4.0"
boto3 = "^1.35.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"